Moved AddQuestsToCrcDatabase to the utility source set, as it's not the kind of utility that holocore would ever use at runtime
